+++ src/pages/home/home.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { HTML5Backend } from 'react-dnd-html5-backend';
+import { DndProvider } from 'react-dnd';
+import BurgerIngredients from '../../components/burger-ingredients/burger-ingredients';
+import BurgerConstructor from '../../components/burger-constructor/burger-constructor';
+
+export const Home = () => {
+  return (
+    <DndProvider backend={HTML5Backend}>
+      <BurgerIngredients />
+      <BurgerConstructor />
+    </DndProvider>
+  )
+}
