+++ src/pages/index.js
@@ -1,11 +1,23 @@
-import { RegisterPage } from "./register-page/register-page"
-import { LoginPage } from "./login-page/login-page"
-import { ForgotPasswordPage } from "./forgot-pass-page/forgot-pass-page"
-import { ResetPasswordPage } from "./reset-pass-page/reset-pass-page"
-import { ProfilePage } from "./profile-page/profile-page"
-import { Error404Page } from "./error-404-page/error-404-page"
-import { IngredientPage } from "./ingredient-page/ingredient-page"
+import { RegisterPage } from "./register/register"
+import { LoginPage } from "./login/login"
+import { ForgotPasswordPage } from "./forgot-pass/forgot-pass"
+import { ResetPasswordPage } from "./reset-pass/reset-pass"
+import { ProfilePage } from "./profile/profile"
+import { Error404Page } from "./error-404/error-404"
+import { IngredientPage } from "./ingredient/ingredient"
 import { OrdersList } from "./orders-list/orders-list"
 import { OrdersHistory } from "./orders-history/orders-history"
+import { Home } from "./home/home"
 
-export {RegisterPage, LoginPage, ForgotPasswordPage, ResetPasswordPage, ProfilePage, Error404Page, IngredientPage, OrdersList, OrdersHistory }
+export { 
+  RegisterPage, 
+  LoginPage, 
+  ForgotPasswordPage, 
+  ResetPasswordPage, 
+  ProfilePage, 
+  Error404Page, 
+  IngredientPage, 
+  OrdersList, 
+  OrdersHistory,
+  Home
+}
