+++ src/pages/login/login.jsx
@@ -1,19 +1,24 @@
 import React, { useRef } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { Link } from 'react-router-dom'
+import { useDispatch, useSelector } from 'react-redux'
+import { Link, Redirect, useLocation } from 'react-router-dom'
 import { Button, Input } from '@ya.praktikum/react-developer-burger-ui-components'
+import { logIn, setFormValue } from '../../services/actions/auth'
 
-import { setFormValue, createUser } from '../../services/actions/auth'
+import styles from './login.module.css'
 
-import styles from './register-page.module.css'
-
-export const RegisterPage = () => {
+export const LoginPage = () => {
 
   const {
-    nameValue,
     emailValue,
-    passwordValue
-  } = useSelector(store => store.formDataReducer.form)
+    passwordValue,
+    user
+  } = useSelector(store => ({
+    emailValue: store.formDataReducer.form.emailValue,
+    passwordValue: store.formDataReducer.form.passwordValue,
+    user: store.userReducer.user
+  }))
+
+  const location = useLocation()
 
   const inputRef = useRef(null)
   const dispatch = useDispatch()
@@ -27,7 +32,12 @@ export const RegisterPage = () => {
   }
 
   const submitForm = () => {
-    dispatch(createUser(nameValue, emailValue, passwordValue))
+    dispatch(logIn(emailValue, passwordValue))
+  }
+
+  if (user) {
+    const { from } = location.state || {from: '/'}
+    return <Redirect to={from} />
   }
 
   return (
@@ -35,22 +45,10 @@ export const RegisterPage = () => {
 
       <div className={styles.form_title}>
         <h2 className="text text_type_main-medium">
-          Регистрация
+          Вход
         </h2>
       </div>
       <div className={styles.form_body}>
-        <Input
-          type={'text'}
-          placeholder={'Имя'}
-          onChange={onFormChange}
-          value={nameValue}
-          name={'nameValue'}
-          error={false}
-          ref={inputRef}
-          onIconClick={onIconClick}
-          errorText={'Ошибка'}
-          size={'default'}
-        />
         <Input
           type={'email'}
           placeholder={'E-mail'}
@@ -77,13 +75,17 @@ export const RegisterPage = () => {
           size={'default'}
         />
         <Button onClick={submitForm} type="primary" size="large">
-          Зарегистрироваться
+          Войти
         </Button>
       </div>
       <div className={styles.form_links}>
         <p className="text text_type_main-default text_color_inactive">
-          Уже зарегистрированы? &nbsp;
-          <Link className={styles.form_links_item} to='/login'>Войти</Link>
+          Вы — новый пользователь? &nbsp;
+          <Link className={styles.form_links_item} to='/register'>Зарегистрироваться</Link>
+        </p>
+        <p className="text text_type_main-default text_color_inactive">
+          Забыли пароль? &nbsp;
+          <Link className={styles.form_links_item} to='/forgot-password'>Восстановить пароль</Link>
         </p>
       </div>
     </div>
